% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rnaseq.R
\name{launch_app_rnaseq}
\alias{launch_app_rnaseq}
\title{Run the shiny application}
\usage{
launch_app_rnaseq(master = "sc://172.18.0.1:15002", ...)
}
\arguments{
\item{...}{further arguments that can be passed to \code{\link[shiny]{shinyApp}} options}
}
\value{
shinyApp object, see also \code{\link[shiny]{shinyApp}}
}
\description{
Run the shiny application
}
