% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq-package.R
\docType{package}
\name{rnaseq-package}
\alias{rnaseq}
\alias{rnaseq-package}
\title{rnaseq: Application to demonstrate HeatMap of RNA Seq}
\description{
Application to demonstrate HeatMap of RNA Seq.
}
\keyword{internal}
