% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movedesign-package.R
\docType{package}
\name{shinyapps-package}
\alias{shinyapps}
\alias{shinyapps-package}
\title{shinyapps: PKG_TITLE}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

PKG_DESC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecoisilva/shinyapps}
  \item Report bugs at \url{https://github.com/ecoisilva/shinyapps/issues}
}

}
\author{
\strong{Maintainer}: AUTHOR_FIRST AUTHOR_LAST \email{AUTHOR@MAIL.COM}

}
\keyword{internal}
